+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Introduction of a new module.
 ## mini.test
 
 - BREAKING FEATURE: `child.get_screenshot()` now by default calls `:redraw` prior to computing screenshot. Can be disabled by new `opts.redraw` argument.
+- FEATURE: new method `child.lua_func()` can execute simple functions inside child process and return the result (stasjok, #437).
 
 
 # Version 0.10.0